Complete failed refactoring
