+++ src/main/java/helltech/dynamodb/model/Publication.java
@@ -7,7 +7,7 @@ import static helltech.dynamodb.DatabaseConstants.PK3;
 import static helltech.dynamodb.DatabaseConstants.SK2;
 import static helltech.dynamodb.DatabaseConstants.SK3;
 import static software.amazon.awssdk.enhanced.dynamodb.TableSchema.fromBean;
-import helltech.dynamodb.Generated;
+import helltech.dynamodb.annotations.Generated;
 import java.util.Objects;
 import java.util.UUID;
 import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
